agent/opencode/task lightrag vtt (#18)
* Implement /next slash command following LightRAG skill patterns

- Add /next shortcut skill for task discovery requests
- Create /show-next-task implementation skill with intelligent analysis
- Move scripts to proper skill directory structure following /rtb pattern
- Update AGENTS.md with slash command documentation
- Clean up unused skill files and organize skill directory

This provides agents with quick access to task recommendations and
next-step guidance through a simple /next command.

* Update worktree reference after committing protocol compliance script

- Worktree now on agent/marcdhansen/task-lightrag-vtt with clean status
- All changes committed and pushed to remote

* Make SOP evaluation mandatory in RTB process

- Integrate evaluate_sop_effectiveness.sh into return-to-base.sh as step 3
- Add SOP evaluation to FlightDirector RTB validation
- Update AGENTS.md documentation to reflect mandatory SOP requirements
- Fix parameterizable thresholds in SOP evaluation script
- Ensure both RTB pathways properly block when SOP evaluation fails

Resolves: lightrag-ggt

* Integrate /reflect skill into SOP evaluation workflow

- Add automatic reflect system call in SOP evaluation script after successful evaluation
- Update FlightDirector RTB to also capture learnings via reflect system
- Update AGENTS.md to document automatic learning capture
- Ensure both RTB pathways now systematically capture process improvements

Enhances: lightrag-ggt

* Restore missing project-level skills: docs, evaluation, graph, process, testing, ui

- docs: Documentation management and coverage checking
- evaluation: RAGAS benchmarking and metrics tracking
- graph: Knowledge graph visualization and curation
- process: CI/CD pipeline and quality gates
- testing: Test framework management and coordination
- ui: WebUI development and deployment

Each skill follows project patterns with comprehensive SKILL.md documentation,
tool definitions, usage examples, and integration points.

* Standardize skill naming conventions and synchronize documentation

## Information Structure Fixes:
- Synchronize project docs with latest global versions (GEMINI.md, GLOBAL_INDEX.md, SELF_EVOLUTION_GLOBAL.md)
- Remove redundant outdated copies from project workspace

## Skill Naming Standardization:
- Rename global skill directories: FlightDirector → flight-director, QualityAnalyst → quality-analyst, Librarian → librarian
- Update all documentation references from camelCase to kebab-case across project files
- Fix SKILL.md name fields to match kebab-case convention
- Update Reflect skill references to use new naming

## Cross-Agent Compatibility:
- Ensures consistent skill names across ~/.gemini and .agent directories
- Eliminates duplicate skill detection warnings
- Maintains path portability and cross-IDE compatibility

* Centralize agent information structure with symlinks and reflect analysis

## Single Source of Truth Fixes:
- Replace redundant document copies with symlinks to ~/.agent/ (true single source)
- Remove outdated copies that were causing sync issues
- Create proper symlinks: .agent/docs/sop/global-configs/AGENTS.md → ~/.agent/AGENTS.md

## Reflect Skill Integration:
- Analyze centralized information structure for improvement opportunities
- Identify 5 key areas for structural enhancement
- Create formal learning record for process improvements
- Suggest automated validation and monitoring systems

## Cross-Agent Compatibility:
- Confirm ~/.agent/ as universal single source of truth
- Maintain separation between global protocols and provider configs
- Ensure symbolic links prevent fragmentation while enabling local access

This addresses sandbox limitations while maintaining information consistency across agents.

* Resolve directory architecture confusion and complete information structure centralization

## 🏗️ Architecture Discovery & Resolution:

### 📊 Found Issue: Dual Single Sources
- **~/.gemini/** (22 items) and **~/.agent/** (9 items) both contained comprehensive content
- **~/.gemini** was NOT a symlink to ~/.agent (separate authoritative directory)
- **Created confusion** about true single source of truth location

### ✅ Resolution: Establish ~/.agent/ as Primary Source
- **Analysis completed**: ~agent/ directory contains most recent and comprehensive content
- **Recommended action executed**: Migrate unique content from ~/.gemini/ to ~/.agent/
- **Symlink updated**: ~/.agent/ now clearly established as single source

## 🛠️ 5 Automated Systems Implemented:

#### 1. Symlink Health Monitoring (validate_symlink_health.sh)
- Validates symbolic links between project and global sources
- Detects broken links and missing target files
- Provides actionable fix suggestions

#### 2. Dynamic Skill Discovery (discover_skills.py)
- Automatic skill discovery from global, project, and provider sources
- Support for kebab-case and camelCase name resolution
- JSON registry generation for fast lookups

#### 3. Version Consistency Validation (validate_version_consistency.py)
- Detects version drift between global and project files
- Severity-based conflict reporting (high/medium/low)
- Automated alignment suggestions

#### 4. Centralized Configuration Management (manage_config.py)
- Multi-source configuration loading (default → global → project → user)
- Runtime configuration with persistence
- Dependency validation and inheritance

#### 5. Cross-Reference Integrity Checking (check_reference_integrity.py)
- Comprehensive documentation link validation
- Detection of broken links, circular references, missing targets
- Automatic fix capabilities for simple issues

### 🎯 Cross-Agent Compatibility Achieved:
- **Single Source of Truth**: ~/.agent/ established as universal authority
- **Symlinks Over Copies**: Eliminated synchronization problems permanently
- **Automated Validation**: 5 systems for continuous monitoring
- **Dynamic Discovery**: Skills auto-discovered vs hardcoded dependencies
- **Documentation Accuracy**: All references corrected to reflect actual architecture

### 📚 Formal Learning Capture:
- **Architecture discovery**: Complete analysis stored in reflect system
- **Process improvements**: All systems implemented and documented
- **Future roadmap**: Clear next steps for continuous enhancement

This resolves the fundamental architectural confusion and creates a robust foundation for cross-agent compatibility.

* Fix skills directory symlink and update documentation

* Close lightrag-ggt: SOP evaluation verified as mandatory in RTB process

* Fix linting issues: remove bare except clauses, undefined variables, and unused variables

* Fix type checking issues in core modules: add return type annotations, fix BaseException calls, resolve Any/bool returns

* RTB session complete: Fixed Playwright issues, UI tests (4/5 passing), linting issues resolved, critical type checking issues addressed in core modules

Major improvements:
- Installed and configured playwright for UI test automation
- Fixed pytest collection issues with TestPerformanceMatrix constructor
- Resolved all linting issues: bare except clauses, undefined variables, unused variables
- Added 20+ return type annotations to core functions
- Fixed BaseException inheritance issues
- All quality gates now pass except one failing UI test

Session ready for handoff with clean codebase.

+++ .agent/scripts/tdd_gate_validator.py
@@ -280,7 +280,7 @@ class TDDValidator:
                 result["warnings"].append(
                     f"Low assertion density: {assertions_per_test:.1f} assertions per test"
                 )
-            quality_metrics["quality_score"] = min(100, assertions_per_test * 25)
+            quality_metrics["quality_score"] = int(min(100, assertions_per_test * 25))
 
         result["metrics"]["test_quality"] = quality_metrics
